// lib/prettier/lib/types/enums.ts
export enum EArrowParens {
  Always = "always",
  Avoid = "avoid",
}

export enum EEmbeddedLanguageFormatting {
  Auto = "auto",
  Off = "off",
}

export enum EEndOfLine {
  Lf = "lf",
  CrLf = "crlf",
  Cr = "cr",
  Auto = "auto",
}

export enum EHtmlWhitespaceSensitivity {
  Css = "css",
  Strict = "strict",
  Ignore = "ignore",
}

export enum EProseWrap {
  Always = "always",
  Never = "never",
  Preserve = "preserve",
}

export enum EQuoteProps {
  AsNeeded = "as-needed",
  Consistent = "consistent",
  Preserve = "preserve",
}

export enum ETrailingComma {
  All = "all",
  Es5 = "es5",
  None = "none",
}